Added -Wall etc. moved more code to Utils
