+++ CMakeLists.txt
@@ -3,7 +3,14 @@ project(CPP_GAME LANGUAGES CXX)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
-add_compile_options("$<$<CONFIG:DEBUG>:-DDEBUG>")
+
+add_compile_options(
+    $<$<CONFIG:DEBUG>:-DDEBUG>
+    $<$<CONFIG:DEBUG>:-Wall>
+    $<$<CONFIG:DEBUG>:-Wextra>
+    $<$<CONFIG:DEBUG>:-Wall>
+    $<$<CONFIG:DEBUG>:-Wpedantic>
+)
 
 include(FetchContent)
 FetchContent_Declare(SFML
